Features

+++ app/controllers/car.controller.js
@@ -1,5 +1,8 @@
 const db = require("../models");
+const repairModel = require("../models/repair.model");
+const upgradeModel = require("../models/upgrade.model");
 const Cars = db.car;
+const User = db.user
 
 //Create and save a new Car
 module.exports = {
@@ -7,7 +10,7 @@ async create(req, res, next) {
  
   //Create a new car
   const car = new db.car({
-    user_id: req.body.user_id,
+    user_id: req.params.userId,
     brand: req.body.brand,
     model: req.body.model,
     colour: req.body.colour,
@@ -20,6 +23,13 @@ async create(req, res, next) {
       : true,
     isDailyCar: req.body.isDailyCar ? req.body.isDailyCar : true,
   });
+
+  const user = await User.findOne({_id : req.params.userId})
+  console.log(user)
+
+  if (user == null || user == undefined) {
+    res.status(404).send("User does not exist")
+  } else {
 
   //Save car in the database
   await car
@@ -34,6 +44,21 @@ async create(req, res, next) {
           "Some error occured while adding the car to the database!",
       });
     });
+
+    try {
+      await User.updateOne(
+        {_id : req.params.userId},
+        {
+          $push: {
+            cars : car,
+          },
+        },
+        {runValidators : true}
+      )
+    } catch (err) {
+      res.status(500).send("Something went wrong");
+    }
+  }
 },
 
 //Retreive all Cars from the database
@@ -54,7 +79,7 @@ async findAll(req, res, next){
 
 //Find a car by ID
 async findOne(req, res, next){
-  const id = req.params.id;
+  const id = req.params.carId;
 
   await Cars.findById(id)
     .then((data) => {
@@ -73,13 +98,23 @@ async findOne(req, res, next){
 
 //Update a car by ID
 async update(req, res, next){
+  const userId = req.params.userId;
+  const id = req.params.carId;
+  const car = await Cars.findOne({_id : id})
+
+  console.log(userId)
+  console.log(id)
+  console.log(car.user_id)
+
+  if (userId != car.user_id) {
+    return res.status(403).send("User is not allowed access");
+  }
+
   if (!req.body) {
     return res.status(400).send({
       message: "Data to update can not be empty. Nothing to update!",
     });
   }
-
-  const id = req.params.id;
 
   await Cars.findByIdAndUpdate(id, req.body, { useFindAndModify: false })
     .then((data) => {
@@ -94,11 +129,33 @@ async update(req, res, next){
         message: "Error updating car with id " + id + "in database!",
       });
     });
+
+    //TODO Fix this
+    // try {
+    //   await User.updateOne(
+    //     { _id : userId },
+    //     {
+    //       $set : {
+    //         cars : req.body
+    //       },
+    //     },
+    //     { runValidators : true }
+    //   )
+    // } catch (err) {
+    //   res.status(500).send("Something went wrong")
+    // }
 },
 
 //Delete a car by ID
 async delete(req, res,next){
-  const id = req.params.id;
+  const id = req.params.carId;
+  const userId = req.params.userId;
+  const car = await Cars.findOne({ _id : id})
+
+
+  if (userId != car.user_id) {
+    return res.status(403).send("User is not allowed access");
+  }
 
   await Cars.findByIdAndRemove(id)
     .then((data) => {
@@ -120,5 +177,18 @@ async delete(req, res,next){
           "! This is a problem with the database connection!",
       });
     });
+    try {
+      await User.updateOne(
+        { _id: userId},
+        {
+          $pull: {
+            Cars: req.body,
+          },
+        },
+        { runValidators: true}
+      )
+    } catch (err) {
+      res.status(500).send("Something went wrong in the query.")
+    }
   }
 }
